Add files via upload
The file must be saved as an “Excel Macro-Enabled Workbook”, or the code will not run. The function of this file is to outline the process of assigning technological licenses to employees and users, and to automate it to the greatest extent possible. Using these macros within the workbook will modify user and license allocation, while simultaneously updating inventory figures within the configured data table.